#321 - Implement http client
The http client uses PHP streams to send requests and provides for a
generic send() method and specific get(), post(), put() and delete()
method functions.

The generic send method will accept a request object and return a
response object, while the specific method function accept a url,
request header array and for none-safe methods data to send and return
any content they receive as part of the response as an array.

If the data is an array it will be send as x-www-form-urlencoded, if
it's a KObjectConfigFormat, it will be send using the specific format
media type.