#360 - Fix module params handling
Module params are passed through a html attribye and need to be escaped
for xml compliance and decoded before send the json string to Joomla for
further processing.

Improve the json() template method to allow defining if the output should
be escaped or not. Default the result is not escaped.