#475: Database layer refactor
Add KDatabase::getAdapter that returns an application-wide singleton for the app database adapter
Add KDatabaseAdapterAbstract::getQuery($identifier) to return a query object with the adapter set
Add KDatabase::getQuery($identifier) as a shortcut to get a query with the app database adapter set
Set the default adapter with KDatabase::getAdapter in KDatabaseTableAbstract
Remove the KObjectMultiton from KDatabaseAdapterAbstract. Now that we have KDatabase::getAdapter() no need to make every adapter a singleton especially since with SQLite there will be a need for different adapters with the same identifier.
Add KObjectMultiton to ComKoowaDatabaseAdapterMysqli so that we are backwards compatible with code that directly calls the adapter with the identifier like: $this->getObject('database.adapter.mysqli');
Add KDatabaseAdapterPdo that accepts any PDO connection string
Add ComKoowaDatabaseAdapterPdo that knows how to reuse the Joomla PDO adapter
Add KDatabaseAdapterSqlite as a specialised PDO adapter that knows how to parse SQLite table information